+++ src/ru/kanban/task/Task.java
@@ -3,19 +3,19 @@ package ru.kanban.task;
 import java.util.Objects;
 
 public class Task {
-    protected long id;
+    protected int id;
     protected String title;
     protected String description;
     protected TaskStatus status;
 
-    public Task(long id, String title, String description) {
+    public Task(int id, String title, String description) {
         this.id = id;
         this.title = title;
         this.description = description;
         this.status = TaskStatus.NEW;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
